pyproject.toml
[project]
name = "moniker-svc"
version = "0.1.0"
description = "Unified data access layer with hierarchical ownership and telemetry"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",  # For REST adapter
]

[project.optional-dependencies]
# Database adapters
snowflake = ["snowflake-connector-python>=3.6.0"]
oracle = ["oracledb>=2.0.0"]

# File formats
excel = ["openpyxl>=3.1.0"]
parquet = ["pyarrow>=14.0.0"]

# Telemetry sinks
zmq = ["pyzmq>=25.1.0"]

# Financial data providers (require commercial licenses)
bloomberg = ["blpapi>=3.19.0"]
refinitiv = ["eikon>=1.1.0", "refinitiv-data>=1.5.0"]

# Development
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pyyaml>=6.0.0",
]

# All optional adapters (except commercial)
all = [
    "snowflake-connector-python>=3.6.0",
    "oracledb>=2.0.0",
    "openpyxl>=3.1.0",
    "pyarrow>=14.0.0",
    "pyzmq>=25.1.0",
]

[project.scripts]
moniker-svc = "moniker_svc.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moniker_svc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "blpapi.*",
    "eikon.*",
    "refinitiv.*",
    "snowflake.*",
    "oracledb.*",
]
ignore_missing_imports = true
